requirements.txt
Pillow==7.0.0
python-dateutil==2.7.3
tk==0.1.0
